get swig wrappers for tree class working (at least for range queries)

+++ cpp/src/global/type_defs.h
@@ -15,7 +15,7 @@ typedef double data_t;
 //typedef float data_t;
 //typedef int	int_t;
 //typedef short int short_int_t;
-typedef int16_t length_t;		//length of a time series in main memory; needs to be signed
+typedef int32_t length_t;		//length of a time series in main memory; needs to be signed
 typedef int16_t idx_t;			//indices in a time series in main memory; needs to be signed
 typedef long tick_t;			// indices in an arbitrary-length time series; needs to be signed
 typedef unsigned short int steps_t;
@@ -31,8 +31,8 @@ extern const float kFLOATING_PT_ERROR;
 
 /** Struct of index + value at that index */
 typedef struct Index {
-	data_t	value;
-	tick_t    index;
+	data_t value;
+	tick_t index;
 } Index;
 
 #endif
